init appCommon
