--- androidApp/src/main/res/values/styles.xml
@@ -1,3 +1,0 @@
-<resources>
-    <style name="AppTheme" parent="android:Theme.Material.NoActionBar"/>
-</resources>
