+++ build-logic/convention/build.gradle.kts
@@ -1,0 +1,5 @@
+plugins {
+    `kotlin-dsl`
+}
+
+group = "com.stslex.atten.buildlogic"
