+++ commonApp/src/iosMain/kotlin/com/stslex/atten/MainViewController.kt
@@ -1,0 +1,5 @@
+package com.stslex.atten
+
+import androidx.compose.ui.window.ComposeUIViewController
+
+fun MainViewController() = ComposeUIViewController { App() }
