+++ iosApp/Podfile
@@ -1,5 +1,5 @@
 target 'iosApp' do
   use_frameworks!
   platform :ios, '16.0'
-  pod 'shared', :path => '../shared'
+  pod 'commonApp', :path => '../commonApp'
 end
