--- shared/src/androidMain/kotlin/com/stslex/atten/Platform.android.kt
@@ -1,7 +1,0 @@
-package com.stslex.atten
-
-class AndroidPlatform : Platform {
-    override val name: String = "Android ${android.os.Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()
