--- shared/src/commonMain/kotlin/com/stslex/atten/Greeting.kt
@@ -1,9 +1,0 @@
-package com.stslex.atten
-
-class Greeting {
-    private val platform: Platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
