--- shared/src/commonMain/kotlin/com/stslex/atten/Platform.kt
@@ -1,7 +1,0 @@
-package com.stslex.atten
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
