--- shared/src/iosMain/kotlin/com/stslex/atten/Platform.ios.kt
@@ -1,9 +1,0 @@
-package com.stslex.atten
-
-import platform.UIKit.UIDevice
-
-class IOSPlatform: Platform {
-    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
-
-actual fun getPlatform(): Platform = IOSPlatform()
